feat: refactor kyma condition handling
this reduces the amount of status updates and reconciliation loops by interacting with the "Module" abstraction which combines all module settings, the current component as well as the channel into a struct that is used in the reconciler. Doing this, we can refactor a lot of state handling and only update when its really necessary, not after every change on the status. E.g. now when modules are already up to date or have state changes, we only update once for all conditions and immediately check for their state, saving 1 reconcile loop.