!feat: add semver-version based skew handling
This adds channel skew handliing for semantic version differences in channesl. This is done to catch a case, in which a Kyma instance was created with a Module in a specific version (e.g. 2.1.0), but the module was already installed with a higher version of that module before (e.g. 3.0.0). This can happen if the channel configuration changed.

This is a regular scenario for cases where the user downgrades a previous module installation in the stability (e.g. from fast->stable). This will lead to the module not being downgraded (which is an unsupported versioning path), and instead it will wait for the stable branch to catchup.

If the user wants to completely downgrade a module instead (e.g. because of bugs), he will have to first uninstall a module, and then reinstall the module in the desired channel (e.g. remove the module from kyma in fast channel, and then add it again in stable). This will cause a reinstallation.

It also removes Observed Generation, as it was never used or referenced in checks since we switched to ModuleTemplate Parsing + Remote Sync on every Reconciliation.