!feat: SSA for module apply / fix namespace handling / fix test data for credSecretSelector (#363)
!feat: SSA for module apply

This PR introduces concurrent SSA for the Manifests generated from the ModuleTemplates and the Kyma CR. Due to this change we also no longer need the "statusUpdateRequired" pattern.

In a follow-up PR it is easily possible now to update the Status Update to SSA as well to decrease possible optimistic lock problems.

It also fixes up the Module Template Test that was touched by https://github.com/kyma-project/lifecycle-manager/pull/358. The test data was changed here to not have a config layer anymore and thus caused issues as the test case was always expecting a config layer. thus the test is now changed to allow empty config layers.